Core(fix): fix the process of ending an IRC session
Server is not always reply a QUIT message when it received a QUIT
message. But it *MUST* close the connection to a client which sends a QUIT
message[1].

So, SRN_SERVER_STATE_CONNECTED + SRN_SERVER_ACTION_QUIT now lead to sending a
QUIT message. and server will be free when its connection is closed by
remote.

Now srain can quit normally from ngircd :D .

[1]: https://tools.ietf.org/html/rfc1459#section-4.1.6

ref: #99