Sui(improve): let GtkMenuButton use default none direction icon
The GtkMenuButton widget can hold any valid child widget.
That is, it can hold almost any other standard GtkWidget.
The most commonly used child is GtkImage. If no widget is
explicitely added to the GtkMenuButton, a GtkImage is
automatically created, using an arrow image oriented
according to “direction” or the generic “open-menu-symbolic”
icon if the direction is not set.