Add support for IRCv3 standard-replies
Srain doesn't currently use any of the specs that depend on standard-replies,
but Ergo sometimes emits them in addition to/instead of numerics (typically
when trying to express an error that doesn't have a numeric).

This is mostly just to suppress the error
`_sirc_event_hdr: assertion 'events->unknown' failed` when failing SASL
authentication on Ergo.

https://ircv3.net/specs/extensions/standard-replies